chore: Implement Webhook Event Discovery
