--- app/Filament/Resources/WebhookConfigurationResource.php
@@ -1,66 +1,0 @@
-<?php
-
-namespace App\Filament\Resources;
-
-use App\Filament\Resources\WebhookConfigurationResource\Pages;
-use App\Models\WebhookConfiguration;
-use App\Services\Webhooks\DiscoverWebhookEventsService;
-use Filament\Forms;
-use Filament\Forms\Form;
-use Filament\Resources\Resource;
-use Filament\Tables;
-use Filament\Tables\Table;
-
-class WebhookConfigurationResource extends Resource
-{
-    protected static ?string $model = WebhookConfiguration::class;
-
-    protected static ?string $navigationIcon = 'tabler-webhook';
-
-    protected static ?string $navigationGroup = 'Webhooks';
-
-    public static function form(Form $form): Form
-    {
-        return $form
-            ->schema([
-                Forms\Components\TextInput::make('endpoint')->activeUrl()->required(),
-                Forms\Components\CheckboxList::make('events')->options(fn () => DiscoverWebhookEventsService::toFilamentCheckboxList())->required(),
-                Forms\Components\TextInput::make('description')->nullable(),
-            ]);
-    }
-
-    public static function table(Table $table): Table
-    {
-        return $table
-            ->columns([
-                //
-            ])
-            ->filters([
-                //
-            ])
-            ->actions([
-                Tables\Actions\EditAction::make(),
-            ])
-            ->bulkActions([
-                Tables\Actions\BulkActionGroup::make([
-                    Tables\Actions\DeleteBulkAction::make(),
-                ]),
-            ]);
-    }
-
-    public static function getRelations(): array
-    {
-        return [
-            //
-        ];
-    }
-
-    public static function getPages(): array
-    {
-        return [
-            'index' => Pages\ListWebhookConfigurations::route('/'),
-            'create' => Pages\CreateWebhookConfiguration::route('/create'),
-            'edit' => Pages\EditWebhookConfiguration::route('/{record}/edit'),
-        ];
-    }
-}
