--- app/Filament/Resources/WebhookConfigurationResource/Pages/CreateWebhookConfiguration.php
@@ -1,11 +1,0 @@
-<?php
-
-namespace App\Filament\Resources\WebhookConfigurationResource\Pages;
-
-use App\Filament\Resources\WebhookConfigurationResource;
-use Filament\Resources\Pages\CreateRecord;
-
-class CreateWebhookConfiguration extends CreateRecord
-{
-    protected static string $resource = WebhookConfigurationResource::class;
-}
