+++ app/Filament/Resources/WebhookResource/Pages/CreateWebhookConfiguration.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace App\Filament\Resources\WebhookResource\Pages;
+namespace App\Filament\Resources\WebhookConfigurationResource\Pages;
 
 use App\Filament\Resources\WebhookResource;
 use Filament\Resources\Pages\CreateRecord;
 
-class CreateWebhook extends CreateRecord
+class CreateWebhookConfiguration extends CreateRecord
 {
     protected static string $resource = WebhookResource::class;
 }
