--- app/Filament/Resources/WebhookResource/Pages/EditWebhook.php
@@ -1,19 +1,0 @@
-<?php
-
-namespace App\Filament\Resources\WebhookResource\Pages;
-
-use App\Filament\Resources\WebhookResource;
-use Filament\Actions;
-use Filament\Resources\Pages\EditRecord;
-
-class EditWebhook extends EditRecord
-{
-    protected static string $resource = WebhookResource::class;
-
-    protected function getHeaderActions(): array
-    {
-        return [
-            Actions\DeleteAction::make(),
-        ];
-    }
-}
