+++ app/Models/Webhook.php
@@ -2,12 +2,10 @@
 
 namespace App\Models;
 
-use Illuminate\Database\Eloquent\Concerns\HasUuids;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 
 class Webhook extends Model
 {
     use HasFactory;
-    use HasUuids;
 }
