+++ app/Models/WebhookConfiguration.php
@@ -4,16 +4,12 @@ namespace App\Models;
 
 use App\Events\Event;
 use Illuminate\Database\Eloquent\Builder;
-use Illuminate\Database\Eloquent\Concerns\HasUuids;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 
 class WebhookConfiguration extends Model
 {
     use HasFactory;
-    use HasUuids;
-
-    protected $primaryKey = 'uuid';
 
     protected $fillable = [
         'endpoint',
